+++ src/noahtheduke/splint/rules/lint/not_empty.clj
@@ -4,7 +4,19 @@
 
 (ns ^:no-doc noahtheduke.splint.rules.lint.not-empty
   (:require
-    [noahtheduke.splint.rules :refer [defrule]]))
+    [noahtheduke.splint.config :refer [get-config]]
+    [noahtheduke.splint.rules :refer [defrule]]
+    [noahtheduke.splint.diagnostic :refer [->diagnostic]]))
+
+(defn seq-diagnostic [rule form {:syms [?x]}]
+  (->diagnostic rule form
+                {:message "`seq` is idiomatic, gotta learn to love it."
+                 :replace-form (list 'seq ?x)}))
+
+(defn not-empty-diagnostic [rule form {:syms [?x]}]
+  (->diagnostic rule form
+                {:message "`not-empty` is built-in."
+                 :replace-form (list 'not-empty ?x)}))
 
 (defrule lint/not-empty?
   "`seq` returns `nil` when given an empty collection. `empty?` is implemented as
@@ -15,9 +27,14 @@
   ; bad
   (not (empty? coll))
 
-  ; good
+  ; good (chosen style :seq (default))
   (seq coll)
+
+  ; good (chosen style :not-empty)
+  (not-empty coll)
   "
   {:pattern '(not (empty? ?x))
-   :message "`seq` is idiomatic, gotta learn to love it."
-   :replace '(seq ?x)})
+   :on-match (fn [ctx rule form bindings]
+               (condp = (:chosen-style (get-config ctx rule))
+                 :seq (seq-diagnostic rule form bindings)
+                 :not-empty (not-empty-diagnostic rule form bindings)))})